backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
neo4j==5.19.0
pydantic-settings==2.2.1
python-dotenv==1.0.1
httpx==0.27.0
google-generativeai>=0.8.0
pytest==8.2.2
pytest-asyncio==0.23.7
